change default extension
