+++ packages/editor/src/stores/index.ts
@@ -2,3 +2,4 @@ export * from './codemirror-editor';
 export * from './use-resolved-theme';
 export * from './use-collaborative-editor-mode';
 export * from './use-editor-settings';
+export * from './use-default-extensions';
